elements reminder du jeu finis en desktop + reorganisation des chemins img dans Quizz
