+++ src/components/ElementsReminder/ElementsReminder.css
@@ -1,29 +1,155 @@
-.fas {
-    color:black;
-    height: 100px;
-    width: 50px;
+.fa-long-arrow-alt-right {
+  height: 50px;
+  width: 50px;
+  color: #1a1d26;
+  grid-column: 12/24;
+  grid-row: 3;
+ 
+}
+.fa-long-arrow-alt-right:hover {
+  left:-30px;
+  transform: translateX(30px);
+}
+
+.fa-long-arrow-alt-down {
+  height: 45px;
+  width: 50px;
+  color: #1a1d26;
+  grid-column: 13/24;
+  grid-row: 4;
+  
+}
+.fa-long-arrow-alt-down:hover {
+  transform: translateY(15px);
+}
+.fa-long-arrow-alt-left {
+  height: 50px;
+  width: 50px;
+  color: #1a1d26;
+  grid-column: 12/24;
+  grid-row: 5;
+
+}
+.fa-long-arrow-alt-left:hover {
+  transform: translateX(-30px);
+  
+}
+.fa-long-arrow-alt-up {
+  height: 45px;
+  width: 50px;
+  color: #1a1d26;
+  grid-column: 11/24;
+  grid-row: 4;
+  
+}
+.fa-long-arrow-alt-up:hover {
+  transform: translateY(-15px);
+}
+
+.ElementsReminderContainer {
+  background-color: #cfc7c2;
+  height: 100vh;
+  width: 100vw;
+  display: grid;
+  grid-template-columns: repeat(24, 1fr);
+  grid-template-rows: repeat(24, 1fr);
+  grid-gap: 5px;
+}
+#Cleverness {
+  height: 50px;
+  width: 50px;
+  grid-column: 11 / 24;
+  grid-row: 3;
+}
+#Resistance {
+  height: 50px;
+  width: 50px;
+  grid-column: 13 / 24;
+  grid-row: 3;
+}
+#Speed {
+  height: 50px;
+  width: 50px;
+  grid-column: 13 / 24;
+  grid-row: 5;
+}
+#Strength {
+  height: 50px;
+  width: 50px;
+  grid-column: 11 / 24;
+  grid-row: 5;
+}
+/*Responsive a faire plus tard en paysage, adapter en fonction des autres elements du jeu + display none la navbar*/
+@media screen and (max-width : 500px){
+  .fa-long-arrow-alt-right {
+    height: 22px;
+    width: 14px;
+    color: #1a1d26;
+    grid-column: 12/24;
+    grid-row: 4;
+   
   }
+
+   .fa-long-arrow-alt-down {
+    height: 20px;
+    width: 25px;
+    color: #1a1d26;
+    grid-column: 13/24;
+    grid-row: 5;
+    
+  }
+  
+  .fa-long-arrow-alt-left {
+    height: 22px;
+    width: 14px;
+    color: #1a1d26;
+    grid-column: 12/24;
+    grid-row: 6;
+  
+  }
+ 
+  .fa-long-arrow-alt-up {
+    height: 20px;
+    width: 25px;
+    color: #1a1d26;
+    grid-column: 11/24;
+    grid-row: 5;
+  }
+ 
   
   .ElementsReminderContainer {
+    background-color: #cfc7c2;
+    height: 100vh;
+    width: 100vw;
     display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-auto-rows: minmax(100px, auto);
-    grid-gap: 10px;
+    grid-template-columns: repeat(16, 1fr);
+    grid-template-rows: repeat(24, 1fr);
+    grid-gap: 5px;
   }
   #Cleverness {
-    grid-column: 1 / 3;
-    grid-row: 1;
+    height: 20px;
+    width: 20px;
+    grid-column: 11 / 24;
+    grid-row: 4;
   }
   #Resistance {
-    grid-column: 3 / 3;
-    grid-row: 1;
+    height: 20px;
+    width: 20px;
+    grid-column: 13 / 24;
+    grid-row: 4;
   }
   #Speed {
-    grid-column: 1 / 3;
-    grid-row: 3;
+    height: 20px;
+    width: 20px;
+    grid-column: 13 / 24;
+    grid-row: 6;
   }
   #Strength {
-    grid-column: 3 / 3;
-    grid-row: 3;
-  
+    height: 20px;
+    width: 20px;
+    grid-column: 11 / 24;
+    grid-row: 6;
   }
+
+  
+}
