+++ src/components/Quizz/HeroesCards/Carte_Gamora_Story-min.png
