+++ src/components/Quizz/HeroesCards/Fire.gif
