+++ src/components/pages/Jeux.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import Navbar from '../navbar/Navbar';
-
+import ElementsReminder from '../ElementsReminder/ElementsReminder';
 
 function Jeux() {
     return (
-        <div style={{ background: 'green', height: '100vh', width: '100vw' }}>
+        <div style={{ background: 'beige', height: '100vh', width: '100vw' }}>
             Jeux
+            <ElementsReminder />
         </div>
     )
 }
